Refactor: Move reader/writer interfaces to separate files
